+++ options.go
@@ -53,6 +53,14 @@ func (c *Config) RegisterHandler(fun interface{}) {
 	c.brain.RegisterHandler(fun)
 }
 
+// WithLogger is an option to replace the default logger of a bot.
+func WithLogger(logger *zap.Logger) Module {
+	return func(conf *Config) error {
+		conf.logger = logger
+		return nil
+	}
+}
+
 // WithContext is an option to replace the default context of a bot.
 func WithContext(ctx context.Context) Module {
 	return func(conf *Config) error {